# backend/swagger.yml
...
openapi: '3.0.0'
info:
  version: 1.0.0
  title: OntoWorks Application API
  description: OntoWorks Backend Application API documentation
  contact:
    name: Francisco Oliveira
    email: a78416@alunos.uminho.pt
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://epl.di.uminho.pt:8001/api
    description: Main server
  - url: http://localhost:5000/api
    description: Testing server
...
paths:
...
  "/rdf4j/management/listRepos":
    get:
      tags:
      - RDF4J
      summary: List Repositories
      description: Returns all existing repositories.
      responses:
        '200':
          description: Request success
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    readable:
                      type: object
                      properties:
                        datatype:
                          type: string
                        type:
                          type: string
                        value:
                          type: boolean
                    id:
                      type: object
                      properties:
                        type:
                          type: string
                        value:
                          type: boolean
                    title:
                      type: object
                      properties:
                        type:
                          type: string
                        value:
                          type: boolean
                    uri:
                      type: object
                      properties:
                        type:
                          type: string
                        value:
                          type: boolean
                    writable:
                      type: object
                      properties:
                        datatype:
                          type: string
                        type:
                          type: string
                        value:
                          type: boolean
        '400':
          description: Bad request
  "/rdf4j/management/create":
    post:
      tags:
      - RDF4J
      summary: Create a new Repository
      description: |
        Create a new Repository.
        Creating a repository with a Repository ID that already exists will overwrite the old one,
        likely resulting is data loss.
      requestBody:
        description: Login credentials.
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: ["memory", "memory-rdfs-dt", "native", "native-rdfs-dt"]
                  default: memory
                Repository ID:
                  type: string
                Repository title:
                  type: string
                # Persist:
                #   type: string
                # Sync Delay:
                #   type: string
              # required:
              #   - type
              #   - Repository ID
              #   - Repository title
              example:
                type: "memory-rdfs-dt"
                Repository ID: "test"
                Repository title: "test"
      responses:
        '200':
          description: Create success
        '400':
          description: Bad request
        default:
          description: TODO (already existing repo ID)
  "/rdf4j/management/delete/{repository}":
    delete:
      tags:
      - RDF4J
      summary: Delete a Repository
      # description: Delete a Repository
      parameters:
      - in: path
        required: true
        name: repository
        description: Target Repository
        schema:
          type: string
          example: "test"
      responses:
        '200':
          description: Delete success
        '400':
          description: Bad request (or the repository does not exist)
  "/rdf4j/management/delete/{repository}/statements":
    delete:
      tags:
      - RDF4J
      summary: Clear all Repository statements/triples
      # description: Clear all Repository statements/triples
      parameters:
      - in: path
        required: true
        name: repository
        description: Target Repository
        schema:
          type: string
          example: "test"
      responses:
        '200':
          description: Clear success
        '400':
          description: Bad request (or the repository does not exist)
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
...